Move assets to src directory
